+++ app/gui/integration-test/project-view/tableVisualisation.spec.ts
@@ -1,10 +1,13 @@
-import { test, type Locator, type Page } from '@playwright/test'
+import { test, type BrowserContext, type Locator, type Page } from '@playwright/test'
 import * as actions from './actions'
 import { expect } from './customExpect'
 import { mockExpressionUpdate, mockMethodCallInfo } from './expressionUpdates'
 import { CONTROL_KEY } from './keyboard'
 import * as locate from './locate'
 import { graphNodeByBinding } from './locate'
+import singleColumnDates from './table-vis-json/singleColumnDates.json' assert { type: 'json' }
+import singleColumnDatetimes from './table-vis-json/singleColumnDatetimes.json' assert { type: 'json' }
+import singleColumnTimes from './table-vis-json/singleColumnTimes.json' assert { type: 'json' }
 import { mockVisualizationDataUpdate } from './visualizationUpdates'
 
 /** Prepare the graph for the tests. We add the table type to the `aggregated` node. */
@@ -377,3 +380,156 @@ test('GenericGrid Table Visualisation Test - two column - link on second', async
   await expect(numberWidget).toBeVisible()
   await expect(numberWidget).toHaveValue('2')
 })
+
+/* 
+   These tests pair with the Enso tests found at test/Visualization_Tests/src/Table_Visualisation_Integration_Spec.enso
+   Those tests check the json produced by prepare_visualization matches a baseline
+   These tests check that json data then renders correctly in an AG Grid in the GUI
+   If you change the json API you can regen the reference json by commenting in the line of code in 
+   check_equal in Table_Visualisation_Integration_Spec.enso and running those tests
+   Then run the js prettier
+   Remember to comment the write back out
+*/
+
+test('Datetime test - sorting and copying', async ({ page, context }) => {
+  await loadData(page, singleColumnDatetimes)
+  await expectCellDataToBe(
+    page,
+    'Value',
+    '2025-01-02 12:13:14.123[MET]',
+    '2025-01-01 12:13:14.123[MET]',
+    '2025-01-03 12:13:14.123[MET]',
+  )
+  const value = await getHeaderLocator(page, { colHeaderName: 'Value' })
+  await value.click() // Sort ascending
+  await expectCellDataToBe(
+    page,
+    'Value',
+    '2025-01-01 12:13:14.123[MET]',
+    '2025-01-02 12:13:14.123[MET]',
+    '2025-01-03 12:13:14.123[MET]',
+  )
+  await value.click() // Sort descending
+  await expectCellDataToBe(
+    page,
+    'Value',
+    '2025-01-03 12:13:14.123[MET]',
+    '2025-01-02 12:13:14.123[MET]',
+    '2025-01-01 12:13:14.123[MET]',
+  )
+  await value.click() // remove sort
+  await expectCellDataToBe(
+    page,
+    'Value',
+    '2025-01-02 12:13:14.123[MET]',
+    '2025-01-01 12:13:14.123[MET]',
+    '2025-01-03 12:13:14.123[MET]',
+  )
+  await expectCopyingColumnClipboardToBe(
+    page,
+    context,
+    'Value',
+    0,
+    1,
+    '2025-01-02 12:13:14.123[MET]\r\n2025-01-01 12:13:14.123[MET]',
+  )
+})
+
+test('Date test - sorting and copying', async ({ page, context }) => {
+  await loadData(page, singleColumnDates)
+  await expectCellDataToBe(page, 'Value', '2025-01-02', '2025-01-01', '2025-01-03')
+  const value = await getHeaderLocator(page, { colHeaderName: 'Value' })
+  await value.click({ position: { x: 10, y: 10 } }) // Sort ascending
+  await expectCellDataToBe(page, 'Value', '2025-01-01', '2025-01-02', '2025-01-03')
+  await value.click({ position: { x: 10, y: 10 } }) // Sort descending
+  await expectCellDataToBe(page, 'Value', '2025-01-03', '2025-01-02', '2025-01-01')
+  await value.click({ position: { x: 10, y: 10 } }) // remove sort
+  await expectCellDataToBe(page, 'Value', '2025-01-02', '2025-01-01', '2025-01-03')
+  await expectCopyingColumnClipboardToBe(page, context, 'Value', 0, 1, '2025-01-02\r\n2025-01-01')
+})
+
+test('Time test - sorting and copying', async ({ page, context }) => {
+  await loadData(page, singleColumnTimes)
+  await expectCellDataToBe(page, 'Value', '12:14:14.123004', '12:13:14.123004', '12:15:14.123004')
+  const value = await getHeaderLocator(page, { colHeaderName: 'Value' })
+  await value.click({ position: { x: 10, y: 10 } }) // Sort ascending
+  await expectCellDataToBe(page, 'Value', '12:13:14.123004', '12:14:14.123004', '12:15:14.123004')
+  await value.click({ position: { x: 10, y: 10 } }) // Sort descending
+  await expectCellDataToBe(page, 'Value', '12:15:14.123004', '12:14:14.123004', '12:13:14.123004')
+  await value.click({ position: { x: 10, y: 10 } }) // remove sort
+  await expectCellDataToBe(page, 'Value', '12:14:14.123004', '12:13:14.123004', '12:15:14.123004')
+  await expectCopyingColumnClipboardToBe(
+    page,
+    context,
+    'Value',
+    0,
+    1,
+    '12:14:14.123004\r\n12:13:14.123004',
+  )
+})
+
+async function expectCopyingColumnClipboardToBe(
+  page: Page,
+  context: BrowserContext,
+  columnName: string,
+  startRow: number,
+  endRow: number,
+  expectedClipboardText: string,
+) {
+  await context.grantPermissions(['clipboard-read', 'clipboard-write'])
+  await getCellLocator(page, columnName, startRow).click()
+  await page.keyboard.down('Shift')
+  await getCellLocator(page, columnName, endRow).click()
+  await page.keyboard.up('Shift')
+  await page.keyboard.press(`${CONTROL_KEY}+C`)
+  const expectClipboard = expect.poll(() =>
+    page.evaluate(() => window.navigator.clipboard.readText()),
+  )
+  await expectClipboard.toBe(expectedClipboardText)
+}
+
+async function loadData(page: Page, data: any) {
+  await initGraph(page)
+
+  const aggregatedNode = graphNodeByBinding(page, 'aggregated')
+  await aggregatedNode.click()
+  await page.keyboard.press('Space')
+  const tableVisualization = locate.tableVisualization(page)
+  await expect(tableVisualization).toExist()
+
+  await mockVisualizationDataUpdate(
+    page,
+    'Standard.Visualization.Table.Visualization.prepare_visualization',
+    data,
+  )
+}
+
+export type ColumnLocatorOptions = {
+  colId?: string
+  colHeaderName?: string
+}
+
+/**
+ * Returns a locator for the header cell
+ */
+export function getHeaderLocator(page: Page, options: ColumnLocatorOptions) {
+  if (options.colHeaderName) {
+    return page.getByRole('columnheader', { name: options.colHeaderName })
+  }
+  return page.getByRole('columnheader').and(page.locator(`[col-id="${options.colId}"]`))
+}
+
+/**
+ * Returns a locator for the cell based off colId and rowIndex
+ */
+export function getCellLocator(page: Page, colId: string, rowIndex: number) {
+  const locatorString = `[row-index="${rowIndex}"] [col-id="${colId}"]`
+  return page.locator(locatorString)
+}
+
+// Helper function to check cell values in a column
+async function expectCellDataToBe(page: Page, colId: string, ...expectedValues: string[]) {
+  for (let i = 0; i < expectedValues.length; i++) {
+    expect(await getCellLocator(page, colId, i).textContent()).toBe(expectedValues[i])
+  }
+}