+++ CHANGELOG.md
@@ -44,12 +44,17 @@
 - [Add Tableau Hyper write support][12900]
 - [Support character encodings when reading fixed-width files.][13138]
 - [Initial Microsoft Azure support][13144]
+- [Column inference for fixed-width files.][13240]
+- [Specify fixed-width file layout with a `Vector` or `Column` of
+  widths.][13240]
+- [Support `skip_rows` and `on_empty_field` for fixed-width files.][13240]
 
 [12726]: https://github.com/enso-org/enso/pull/12726
 [12950]: https://github.com/enso-org/enso/pull/12950
 [12900]: https://github.com/enso-org/enso/pull/12900
 [13138]: https://github.com/enso-org/enso/pull/13138
 [13144]: https://github.com/enso-org/enso/pull/13144
+[13240]: https://github.com/enso-org/enso/pull/13240
 
 #### Enso Language & Runtime